wip: Create reducers to crud Goals and adjust route
