--- src/main/webapp/app/modules/goals-objectives/models/goals.ts
@@ -1,14 +1,0 @@
-export type TargetGoals = {
-  idUsuarioCriacao?: number;
-  dataCricao: string;
-  qp: string;
-  qt: string;
-  qpd: string;
-  goals: {
-    id: number;
-  }[];
-};
-
-export type Goals = {
-  id: number;
-};
