+++ src/main/webapp/app/modules/goals-objectives/reducers.ts
@@ -1,9 +1,12 @@
 import metaObjetivo from './reducers/meta-objetivo.reducer';
 import metasLista from './reducers/metas-list.reducer';
 import metas from './reducers/metas.reducer';
+import resources from './reducers/resources.reducer';
+
 const metasObjetivosReducers = {
   metaObjetivo,
   metasLista,
+  resources,
   metas,
 };
